raspberry_pi/requirements.txt
Adafruit-DHT==1.4.0
Adafruit-ADS1x15==1.0.2
RPi.GPIO==0.7.1
picamera==1.13
requests==2.31.0
pyserial==3.5
psutil==5.9.5
opencv-python==4.8.0.76
numpy==1.24.3
smbus2==0.4.2
